# backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "S.A.R.A. API",
        "description": "API para el Sistema Ágil de Rutas y Asignaciones",
        "contact": {
            "email": "support@sara.com"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/rutas": {
            "get": {
                "tags": [
                    "Rutas"
                ],
                "summary": "Listar rutas",
                "operationId": "07f8f66135da1afc0dbdded354b14e48",
                "responses": {
                    "200": {
                        "description": "Lista de rutas",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "type": "integer"
                                            },
                                            "nombre": {
                                                "type": "string"
                                            },
                                            "origen": {
                                                "type": "string"
                                            },
                                            "destino": {
                                                "type": "string"
                                            },
                                            "distancia_km": {
                                                "type": "number"
                                            },
                                            "duracion_estimada_min": {
                                                "type": "integer"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Rutas"
                ],
                "summary": "Crear una nueva ruta",
                "operationId": "7ae787bdba22d5f0225ce1875bc8f0d8",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "nombre",
                                    "origen",
                                    "destino",
                                    "distancia_km",
                                    "duracion_estimada_min"
                                ],
                                "properties": {
                                    "nombre": {
                                        "type": "string",
                                        "example": "Ruta Norte"
                                    },
                                    "origen": {
                                        "type": "string",
                                        "example": "Ciudad A"
                                    },
                                    "destino": {
                                        "type": "string",
                                        "example": "Ciudad B"
                                    },
                                    "distancia_km": {
                                        "type": "number",
                                        "example": 150.5
                                    },
                                    "duracion_estimada_min": {
                                        "type": "integer",
                                        "example": 120
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Ruta creada",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer"
                                        },
                                        "nombre": {
                                            "type": "string"
                                        },
                                        "origen": {
                                            "type": "string"
                                        },
                                        "destino": {
                                            "type": "string"
                                        },
                                        "distancia_km": {
                                            "type": "number"
                                        },
                                        "duracion_estimada_min": {
                                            "type": "integer"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/rutas/{id}": {
            "put": {
                "tags": [
                    "Rutas"
                ],
                "summary": "Actualizar una ruta",
                "operationId": "5fd15592c13b8a0d0310bd8112d2c16d",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "nombre",
                                    "origen",
                                    "destino",
                                    "distancia_km",
                                    "duracion_estimada_min"
                                ],
                                "properties": {
                                    "nombre": {
                                        "type": "string",
                                        "example": "Ruta Norte"
                                    },
                                    "origen": {
                                        "type": "string",
                                        "example": "Ciudad A"
                                    },
                                    "destino": {
                                        "type": "string",
                                        "example": "Ciudad B"
                                    },
                                    "distancia_km": {
                                        "type": "number",
                                        "example": 150.5
                                    },
                                    "duracion_estimada_min": {
                                        "type": "integer",
                                        "example": 120
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Ruta actualizada",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer"
                                        },
                                        "nombre": {
                                            "type": "string"
                                        },
                                        "origen": {
                                            "type": "string"
                                        },
                                        "destino": {
                                            "type": "string"
                                        },
                                        "distancia_km": {
                                            "type": "number"
                                        },
                                        "duracion_estimada_min": {
                                            "type": "integer"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Rutas"
                ],
                "summary": "Eliminar una ruta",
                "operationId": "541c24e02b1a6056f57c938357b5ffe3",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Ruta eliminada"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/pedido-ruta": {
            "post": {
                "tags": [
                    "Rutas"
                ],
                "summary": "Asignar ruta a un pedido",
                "operationId": "196c2f8d32cc90bfaedfcb04f437b5f1",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "pedido_id",
                                    "ruta_id"
                                ],
                                "properties": {
                                    "pedido_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "ruta_id": {
                                        "type": "integer",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Ruta asignada",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer"
                                        },
                                        "pedido_id": {
                                            "type": "integer"
                                        },
                                        "ruta_id": {
                                            "type": "integer"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/pedido-ruta/{id}": {
            "delete": {
                "tags": [
                    "Rutas"
                ],
                "summary": "Desasignar ruta de un pedido",
                "operationId": "c18e5b8ec880e4caeafca9ab811aa261",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Ruta desasignada"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
...
}